Fixed a bug and moved updating points to the model
Fixed a bug where CSL's were updating in a broken way, removing their
skillOptions. Gosh, that code is some of the first I wrote and it's a
giant mess...would be nice to go in and clean that up eventually. Also
saw that I was updating the character's spent points in the router, and
moved that to the model. I think that should happen as part of the
update functions for different models.

https://app.asana.com/0/25694918946217/38540345758394
https://app.asana.com/0/25694918946217/38540345758396